+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -68,6 +68,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed the water system not working on metal.
 - Fixed the missing debug menus to visualize the ray tracing acceleration structure (case 1371410).
 - Fixed compilation issue related to shader stripping in ray tracing.
+- Fixed timeline not updating PBR HDAdditionalLightData parameters properly.
 
 ### Changed
 - Converted most TGA textures files to TIF to reduce the size of HDRP material samples.